feat: refactor use-on-click-outside hooks
