+++ src/renderer/src/hooks/index.ts
@@ -1,7 +1,6 @@
 export * from "./use-download";
 export * from "./use-library";
 export * from "./use-date";
-export * from "./use-event-listener";
 export * from "./use-on-click-outside";
 export * from "./use-toast";
 export * from "./redux";
